Refactoring test_launcher into a dedicated module
